01-javase/javase-01-base/src/test/java/cn/gaohank/program/javase/base/base_03_class/Class04InnerTest.java
package cn.gaohank.program.javase.base.base_03_class;

import org.junit.Test;

import static org.junit.Assert.*;

public class Class04InnerTest {
    @Test
    public void testInnerClass() {
        Class04Inner class04Inner = new Class04Inner();
        Class04Inner.InnerClass innerClass = class04Inner.new InnerClass();
        innerClass.setOuterValue(10);
        innerClass.setValue(100);
        assertEquals(10, class04Inner.getValue());
        assertEquals(100, innerClass.getValue());
    }

    @Test
    public void testInnerStaticClass() {
        Class04Inner.InnerStaticClass innerStaticClass = new Class04Inner.InnerStaticClass();
        innerStaticClass.setValue(10);
        assertEquals(10, innerStaticClass.getValue());
    }
}